Added aggregator for throughput values + a number of bug fixes

+++ src/main/scala/rhyskeepence/model/DataPoint.scala
@@ -2,4 +2,4 @@ package rhyskeepence.model
 
 case class DataPoint(timestamp: Long, environment: String, metrics: List[Metric])
 
-case class Metric(name: String, value: Long)
+case class Metric(name: String, value: Double)
